+++ packages/dev/quicksand-gh-runners/quicksand_gh_runners/config.py
@@ -14,8 +14,3 @@ RUNNERS = {
 
 DEFAULT_SHUTDOWN_TIME = "0200"  # 9pm EST = 02:00 UTC
 SHUTDOWN_TIMEZONE = "UTC"
-
-# Azure DevOps
-ADO_ORG = os.environ.get("ADO_ORG", "")
-ADO_FEED = os.environ.get("ADO_FEED", "")
-ADO_SCOPE = "499b84ac-1321-427f-aa17-267ca6975798/.default"